+++ packages/koot/ReactApp/client/run.js
@@ -1,5 +1,4 @@
 // import thunk from 'redux-thunk'
-// import browserHistory from 'react-router/lib/browserHistory'
 // import { routerMiddleware } from 'react-router-redux'
 
 //